LMS1-93 <create update delete display leave requested on admin>

+++ backend_iden/database/migrations/2022_05_01_060321_add_profile_to_users_table.php
@@ -13,9 +13,12 @@ class AddProfileToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('profile')->after('email')->nullable();
-        });
+        // Check if the 'profile' column does not exist before adding it
+        if (!Schema::hasColumn('employees', 'profile')) {
+            Schema::table('employees', function (Blueprint $table) {
+                $table->string('profile')->nullable()->after('email');
+            });
+        }
     }
 
     /**
@@ -25,8 +28,11 @@ class AddProfileToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('profile');
+        Schema::table('employees', function (Blueprint $table) {
+            // Drop the column only if it exists
+            if (Schema::hasColumn('employees', 'profile')) {
+                $table->dropColumn('profile');
+            }
         });
     }
 }
